Professor oak class into header
